server/requirements.txt
asgiref==3.8.1
Django==5.2.1
sqlparse==0.5.3
celery==5.3.6
requests==2.31.0
redis==5.0.4
djangorestframework==3.15.1
django-prometheus==2.3.1
psycopg2-binary
djangorestframework-simplejwt
django-cors-headers==4.6.0
django-celery-results
django-celery-beat

